pyproject.toml
[tool.poetry]
name = "nonebot-adapter-onebot"
version = "2.4.5"
description = "OneBot(CQHTTP) adapter for nonebot2"
authors = ["yanyongyu <yyy@nonebot.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://onebot.adapters.nonebot.dev/"
repository = "https://github.com/nonebot/adapter-onebot"
documentation = "https://onebot.adapters.nonebot.dev/"
keywords = ["bot", "qq", "qqbot", "coolq", "onebot", "cqhttp"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Robot Framework",
  "Framework :: Robot Framework :: Library",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
]
packages = [{ include = "nonebot" }]

[tool.poetry.dependencies]
python = "^3.9"
anyio = "^4.2.0"
sniffio = "^1.3.1"
msgpack = "^1.0.3"
nonebot2 = "^2.2.0"
typing-extensions = ">=4.0.0,<5.0.0"
pydantic = ">=1.10.0,<3.0.0,!=2.5.0,!=2.5.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"
isort = "^5.10.1"
black = "^24.0.0"
nonebug = "^0.3.0"
nonemoji = "^0.1.2"
pytest-cov = "^5.0.0"
pre-commit = "^3.2.0"
nb-autodoc = "^1.0.0a5"
nonebot2 = { git = "https://github.com/nonebot/nonebot2.git", branch = "master", extras = [
  "fastapi",
] }

[tool.pytest.ini_options]
addopts = "--cov=nonebot/adapters/onebot --cov-report term-missing"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
exclude = ["packages/**/*"]

select = [
  "F", # Pyflakes
  "W", # pycodestyle warnings
  "E", # pycodestyle errors
  # "I",     # isort
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "TID",   # flake8-tidy-imports
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "PYI021", # docstring-in-stub
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }

typeCheckingMode = "standard"
reportShadowedImports = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
disableBytesTypePromotions = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
